feat: p2tr script path classification & validation
This adds logic to classify `p2tr` script path spend inputs according to
the BIP 341 script validation rules. It also adds validation logic on
the `p2tr` payment class to validate that a given witness stack matches
the witness program commited to in the output script. It does this by
validating the control block and merkle proof for inclusion of the
selected tapscript in the taptree committed to in the witness program.

The new input classification and validation logic is tested against the
following transaction which uses the script path to spend a taproot
input found in the bitcoin signet chain.

https://explorer.bc-2.jp/tx/dda4cb9290415952d93f52518df978d45ae1710e60e711e6786926e99d305ef9?input:1

Lastly, this commit adds a sanity check to the `p2tr` extended
validation to ensure there is not a mismatch between the address and
output script in the event both are specified.

Ticket: BG-37366