feat: export `classify`
It is kind of a step in the wrong direction: the reason it was not
exported upstream is because `classify` is part of the
TransactionBuilder system and will be removed in `bitcoinjs-lib@6`.

Still, some of BitGo (and utxo-lib) code depends on it so we need to
export it at this point.

Issue: BG-37511